third_party/WebKit/LayoutTests/editing/execCommand/format_block/no-visible-content.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<script>
test(() => assert_selection(
  '<div contenteditable>|</div>',
  'FormatBlock h1',
  '<div contenteditable><h1>|<br></h1></div>'),
  'FormatBlock with no content');
</script>
